--- src/components/movie-session/movie-session-card/styles.scss
@@ -1,9 +1,0 @@
-@import "./../../global";
-
-.container {
-    display: inline-block;
-    background-color: $dark;
-    padding: 10px;
-    margin: 10px;
-    color: $text-color-light;
-}
