+++ src/components/movie/movie-details/action-constants.js
@@ -1,0 +1,3 @@
+export const RECEIVE_MOVIE_DETAILS = 'RECEIVE_MOVIE_DETAILS';
+export const REQUEST_MOVIE_DETAILS = 'REQUEST_MOVIE_DETAILS';
+export const FETCH_MOVIE_DETAILS = 'FETCH_MOVIE_DETAILS';
