+++ src/components/movie/movie-details/styles.scss
@@ -1,4 +1,4 @@
-@import "../../global.scss";
+@import "../../global";
 
 .container {
     padding: 15px;
