+++ src/components/movie/movie-list/movie-card/styles.scss
@@ -1,4 +1,4 @@
-@import '../../global.scss';
+@import '../../../global';
 
 .container {
     display: inline-block;
