--- src/reducers/cinemas.js
@@ -1,20 +1,0 @@
-import {
-  REQUEST_CINEMAS,
-  RECEIVE_CINEMAS,
-} from '../actions/cinema-actions';
-
-const cinemas = (state = [], action) => {
-  switch (action.type) {
-    case REQUEST_CINEMAS:
-      return Object.assign({}, state, { isLoading: true });
-    case RECEIVE_CINEMAS:
-      return Object.assign({}, state, {
-        data: action.cinemas,
-        isLoading: false,
-      });
-    default:
-      return state;
-  }
-};
-
-export default cinemas;
